13.1.2-2 baseline
Former-commit-id: c0c1d56f3aa2efe0f6825d70ff9e56407d690836 [formerly c0c1d56f3aa2efe0f6825d70ff9e56407d690836 [formerly 221d0aaf0e19a054387134e72a76a278159bf4d0]]
Former-commit-id: 4ba416aeb32ba04688841ada586ed9d1be6660f4
Former-commit-id: 941337c5538b8031bb492c28cceecbf7759798be

+++ edexOsgi/com.raytheon.uf.edex.database/src/com/raytheon/uf/edex/database/plugin/PluginDao.java
@@ -35,6 +35,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import org.geotools.coverage.grid.GridCoverage2D;
+import org.geotools.geometry.jts.ReferencedEnvelope;
+import org.opengis.referencing.FactoryException;
+import org.opengis.referencing.crs.CoordinateReferenceSystem;
 import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.support.TransactionCallbackWithoutResult;
 
@@ -51,7 +55,13 @@ import com.raytheon.uf.common.datastorage.IDataStore.StoreOp;
 import com.raytheon.uf.common.datastorage.Request;
 import com.raytheon.uf.common.datastorage.StorageException;
 import com.raytheon.uf.common.datastorage.StorageStatus;
+import com.raytheon.uf.common.datastorage.records.ByteDataRecord;
+import com.raytheon.uf.common.datastorage.records.FloatDataRecord;
 import com.raytheon.uf.common.datastorage.records.IDataRecord;
+import com.raytheon.uf.common.datastorage.records.IntegerDataRecord;
+import com.raytheon.uf.common.geospatial.ISpatialEnabled;
+import com.raytheon.uf.common.geospatial.ISpatialObject;
+import com.raytheon.uf.common.geospatial.MapUtil;
 import com.raytheon.uf.common.localization.IPathManager;
 import com.raytheon.uf.common.localization.LocalizationContext;
 import com.raytheon.uf.common.localization.LocalizationContext.LocalizationLevel;
@@ -60,6 +70,8 @@ import com.raytheon.uf.common.localization.LocalizationFile;
 import com.raytheon.uf.common.localization.PathManagerFactory;
 import com.raytheon.uf.common.serialization.SerializationException;
 import com.raytheon.uf.common.serialization.SerializationUtil;
+import com.raytheon.uf.common.spatial.reprojection.DataReprojector;
+import com.raytheon.uf.common.spatial.reprojection.ReferencedDataRecord;
 import com.raytheon.uf.common.time.util.TimeUtil;
 import com.raytheon.uf.common.util.FileUtil;
 import com.raytheon.uf.edex.core.EdexException;
@@ -71,6 +83,10 @@ import com.raytheon.uf.edex.database.purge.PurgeLogger;
 import com.raytheon.uf.edex.database.purge.PurgeRule;
 import com.raytheon.uf.edex.database.purge.PurgeRuleSet;
 import com.raytheon.uf.edex.database.query.DatabaseQuery;
+import com.vividsolutions.jts.geom.Coordinate;
+import com.vividsolutions.jts.geom.Envelope;
+import com.vividsolutions.jts.geom.Geometry;
+import com.vividsolutions.jts.geom.Polygon;
 
 /**
  * Abstract implementation of a Plugin data access object
@@ -86,6 +102,7 @@ import com.raytheon.uf.edex.database.query.DatabaseQuery;
  * 2/6/09       1990       bphillip    Initial creation
  * 6/29/12      #828       dgilling    Force getPurgeRulesForPlugin()
  *                                     to search only COMMON_STATIC.
+ * Oct 10, 2012 1261       djohnson    Add some generics wildcarding.
  * </pre>
  * 
  * @author bphillip
@@ -174,13 +191,14 @@ public abstract class PluginDao extends CoreDao {
         persistToDatabase(records);
     }
 
+    @SuppressWarnings("unchecked")
     public PluginDataObject[] persistToDatabase(PluginDataObject... records) {
-        List<PersistableDataObject> toPersist = new ArrayList<PersistableDataObject>();
+        List<PersistableDataObject<Object>> toPersist = new ArrayList<PersistableDataObject<Object>>();
         for (PluginDataObject record : records) {
             toPersist.add(record);
         }
-        List<PersistableDataObject> duplicates = mergeAll(toPersist);
-        for (PersistableDataObject pdo : duplicates) {
+        List<PersistableDataObject<Object>> duplicates = mergeAll(toPersist);
+        for (PersistableDataObject<Object> pdo : duplicates) {
             logger.info("Discarding duplicate: "
                     + ((PluginDataObject) (pdo)).getDataURI());
             toPersist.remove(pdo);
@@ -442,21 +460,45 @@ public abstract class PluginDao extends CoreDao {
      *             If problems occur while interacting with the data stores
      */
     public void purgeAllData() throws PluginException {
+        purgeAllData(null);
+    }
+
+    /**
+     * Purges all data associated with the productKeys and owning plugin
+     * 
+     * @throws PluginException
+     *             If problems occur while interacting with the data stores
+     */
+    public void purgeAllData(Map<String, String> productsKeys)
+            throws PluginException {
+        boolean purgeHdf5Data = false;
+        try {
+            // Determine if this plugin uses HDF5 to store data
+            purgeHdf5Data = (PluginFactory.getInstance()
+                    .getPluginRecordClass(pluginName).newInstance() instanceof IPersistable);
+        } catch (Exception e) {
+            PurgeLogger.logError(
+                    "Unabled to determine if plugin has HDF5 data to purge",
+                    this.pluginName, e);
+        }
+
         try {
             List<Date> allRefTimes = getRefTimes();
             Map<String, List<String>> filesToDelete = new HashMap<String, List<String>>();
             for (Date d : allRefTimes) {
-                this.purgeDataByRefTime(d, null, true, false, filesToDelete);
+                this.purgeDataByRefTime(d, productsKeys, purgeHdf5Data, false,
+                        filesToDelete);
             }
-            for (String file : filesToDelete.keySet()) {
-                try {
-                    IDataStore ds = DataStoreFactory
-                            .getDataStore(new File(file));
-                    ds.deleteFiles(null);
-                } catch (Exception e) {
-                    PurgeLogger.logError(
-                            "Error occurred purging file: " + file,
-                            this.pluginName, e);
+            if (purgeHdf5Data) {
+                for (String file : filesToDelete.keySet()) {
+                    try {
+                        IDataStore ds = DataStoreFactory.getDataStore(new File(
+                                file));
+                        ds.deleteFiles(null);
+                    } catch (Exception e) {
+                        PurgeLogger.logError("Error occurred purging file: "
+                                + file, this.pluginName, e);
+                    }
                 }
             }
         } catch (Exception e) {
@@ -1054,7 +1096,9 @@ public abstract class PluginDao extends CoreDao {
     /**
      * Purges data from the database for this plugin with the given reference
      * time matching the given productKeys. If refTime is null, will purge all
-     * data associated with the productKeys.
+     * data associated with the productKeys. Hdf5 must be purged separately as
+     * most hdf5 files can't be purged with a single reference time. Use the
+     * passed map to track what needs to be done with hdf5.
      * 
      * @param refTime
      *            The reftime to delete data for. A null will purge all data for
@@ -1426,8 +1470,8 @@ public abstract class PluginDao extends CoreDao {
             return null;
         }
         try {
-            PurgeRuleSet purgeRules = (PurgeRuleSet) SerializationUtil
-                    .jaxbUnmarshalFromXmlFile(defaultRule);
+            PurgeRuleSet purgeRules = SerializationUtil
+                    .jaxbUnmarshalFromXmlFile(PurgeRuleSet.class, defaultRule);
             return purgeRules.getDefaultRules();
         } catch (SerializationException e) {
             PurgeLogger.logError("Error deserializing default purge rule!",
@@ -1535,5 +1579,131 @@ public abstract class PluginDao extends CoreDao {
         dbQuery.addOrder("dataTime.refTime", true);
 
         return (List<PersistableDataObject>) this.queryByCriteria(dbQuery);
+    }
+
+    public double getHDF5Value(PluginDataObject pdo,
+            CoordinateReferenceSystem crs, Coordinate coord,
+            double defaultReturn) throws Exception {
+        IDataStore store = getDataStore((IPersistable) pdo);
+        // TODO a cache would probably be good here
+        double rval = defaultReturn;
+        if (pdo instanceof ISpatialEnabled) {
+            ISpatialObject spat = getSpatialObject(pdo);
+            DataReprojector reprojector = getDataReprojector(store);
+            ReferencedEnvelope nativeEnv = getNativeEnvelope(spat);
+            IDataRecord data = reprojector.getProjectedPoints(pdo.getDataURI(),
+                    spat, nativeEnv, crs, new Coordinate[] { coord });
+            Double res = extractSingle(data);
+            if (res != null) {
+                rval = res;
+            }
+        }
+        return rval;
+    }
+
+    /**
+     * @param record
+     * @param crs
+     *            target crs for projected data
+     * @param envelope
+     *            bounding box in target crs
+     * @return null if envelope is disjoint with data bounds
+     * @throws Exception
+     */
+    public ReferencedDataRecord getProjected(PluginDataObject record,
+            CoordinateReferenceSystem crs, Envelope envelope) throws Exception {
+        ReferencedEnvelope targetEnv = new ReferencedEnvelope(
+                envelope.getMinX(), envelope.getMaxX(), envelope.getMinY(),
+                envelope.getMaxY(), crs);
+        return getProjected(record, targetEnv);
+    }
+
+    /**
+     * @param record
+     * @param crs
+     *            target crs for projected data
+     * @param envelope
+     *            bounding box in target crs
+     * @return null if envelope is disjoint with data bounds
+     * @throws Exception
+     */
+    public GridCoverage2D getProjectedCoverage(PluginDataObject record,
+            CoordinateReferenceSystem crs, Envelope envelope) throws Exception {
+        ReferencedEnvelope targetEnv = new ReferencedEnvelope(
+                envelope.getMinX(), envelope.getMaxX(), envelope.getMinY(),
+                envelope.getMaxY(), crs);
+        return getProjectedCoverage(record, targetEnv);
+    }
+
+    /**
+     * @param record
+     * @param targetEnvelope
+     *            bounding box in target crs
+     * @return null if envelope is disjoint with data bounds
+     * @throws Exception
+     */
+    public ReferencedDataRecord getProjected(PluginDataObject record,
+            ReferencedEnvelope targetEnvelope) throws Exception {
+        ISpatialObject spatial = getSpatialObject(record);
+        IDataStore store = getDataStore((IPersistable) record);
+        DataReprojector reprojector = getDataReprojector(store);
+        ReferencedEnvelope nativeEnvelope = getNativeEnvelope(spatial);
+        return reprojector.getReprojected(record.getDataURI(), spatial,
+                nativeEnvelope, targetEnvelope);
+    }
+
+    /**
+     * @param record
+     * @param targetEnvelope
+     *            bounding box in target crs
+     * @return null if envelope is disjoint with data bounds
+     * @throws Exception
+     */
+    public GridCoverage2D getProjectedCoverage(PluginDataObject record,
+            ReferencedEnvelope envelope) throws Exception {
+        ISpatialObject spatial = getSpatialObject(record);
+        IDataStore store = getDataStore((IPersistable) record);
+        DataReprojector reprojector = getDataReprojector(store);
+        ReferencedEnvelope nativeEnvelope = getNativeEnvelope(spatial);
+        return reprojector.getReprojectedCoverage(record.getDataURI(), spatial,
+                nativeEnvelope, envelope);
+    }
+
+    protected ISpatialObject getSpatialObject(PluginDataObject record)
+            throws Exception {
+        if (record instanceof ISpatialEnabled) {
+            return ((ISpatialEnabled) record).getSpatialObject();
+        } else {
+            throw new Exception(record.getClass() + " is not spatially enabled");
+        }
+    }
+
+    protected DataReprojector getDataReprojector(IDataStore dataStore) {
+        return new DataReprojector(dataStore);
+    }
+
+    protected ReferencedEnvelope getNativeEnvelope(ISpatialObject spatial)
+            throws FactoryException {
+        CoordinateReferenceSystem crs = spatial.getCrs();
+        Geometry geom = spatial.getGeometry();
+        return MapUtil.getBoundingEnvelope(crs, (Polygon) geom);
+    }
+
+    public Double extractSingle(IDataRecord record) {
+        Double rval = null;
+        if (record == null) {
+            return rval;
+        }
+        if (record instanceof ByteDataRecord) {
+            byte[] data = ((ByteDataRecord) record).getByteData();
+            rval = (double) data[0];
+        } else if (record instanceof FloatDataRecord) {
+            float[] data = ((FloatDataRecord) record).getFloatData();
+            rval = (double) data[0];
+        } else if (record instanceof IntegerDataRecord) {
+            int[] data = ((IntegerDataRecord) record).getIntData();
+            rval = (double) data[0];
+        }
+        return rval;
     }
 }
